feat[Draft]: 连线单机添加 nav 菜单项
